split test between xpack-core part and hlrc part
