+++ client/rest-high-level/src/test/java/org/elasticsearch/client/license/LicenseStatusTests.java
@@ -1,0 +1,38 @@
+/*
+ * Licensed to Elasticsearch under one or more contributor
+ * license agreements. See the NOTICE file distributed with
+ * this work for additional information regarding copyright
+ * ownership. Elasticsearch licenses this file to you under
+ * the Apache License, Version 2.0 (the "License"); you may
+ * not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+package org.elasticsearch.client.license;
+
+import org.elasticsearch.test.ESTestCase;
+
+import static org.hamcrest.CoreMatchers.equalTo;
+
+public class LicenseStatusTests extends ESTestCase {
+
+    public void testCompatibility() {
+        final LicenseStatus[] values = LicenseStatus.values();
+        final LicenseStatus[] hlrcValues = LicenseStatus.values();
+
+        assertThat(values.length, equalTo(hlrcValues.length));
+
+        for (LicenseStatus value : values) {
+            final LicenseStatus licenseStatus = LicenseStatus.fromString(value.label());
+            assertThat(licenseStatus.label(), equalTo(value.label()));
+        }
+    }
+}
